added donations to detail screens and make support for landscape

+++ app/src/main/java/com/kiral/charityapp/domain/model/Charity.kt
@@ -13,7 +13,7 @@ data class Charity(
     val description: String,
     val howDonationHelps: String,
     val whyToDonate: String,
-    var raised: Float,
+    var raised: Double,
     val peopleDonated: Int,
     var donorDonated: Double,
     //val donorId: Int,
